EDIT select de TABLA.PHP se refresca por javascript sin refresca página y varios

+++ obras/obras_proy_delete.php
@@ -21,7 +21,7 @@ if ($rs=Drow( 'OBRAS', "ID_OBRA=$id_obra AND $where_c_coste"))
    $id_prod_estudio_costes = $rs["id_prod_estudio_costes"] ;
    $id_subobra_auto = $rs["id_subobra_auto"] ;
   
-   $sql= "DELETE FROM `PRODUCCIONES_DETALLE`  WHERE ID_PRODUCCION=$id_prod_estudio_costes "  ;     // vaciamos la id_produccion_estudio_coste solamente, las otras RV , por seguridad, deben de hacerlo manualmente    
+   $sql= "DELETE FROM `PRODUCCIONES_DETALLE`  WHERE ID_PRODUCCION='$id_prod_estudio_costes' "  ;     // vaciamos la id_produccion_estudio_coste solamente, las otras RV , por seguridad, deben de hacerlo manualmente    
    $Conn->query($sql) ;
   
 //   $solo_udo_vacias_where= $solo_udo_vacias ? " AND MED_PROYECTO=0 " : ""  ;   
@@ -40,14 +40,39 @@ if ($rs=Drow( 'OBRAS', "ID_OBRA=$id_obra AND $where_c_coste"))
 
     }else
     {
+        
        $sql= "DELETE FROM `Udos`  WHERE ID_OBRA=$id_obra "  ;     // ejecuto DELETE
-       $Conn->query($sql) ;
+       if (!($Conn->query($sql)))   // si hay errores vamos a borrar una a una las UDOS 
+       {
+           $array_udos= DArray("ID_UDO", "Udos", "ID_OBRA='$id_obra'" ) ;
+           foreach ($array_udos as $id_udo) {
+                      $sql= "DELETE FROM `Udos`  WHERE ID_UDO='$id_udo[0]'"  ;     // ejecuto DELETE
+                      $Conn->query($sql) ;
+           }
+
+       }    
 
        $sql= "DELETE FROM `Capitulos`  WHERE ID_OBRA=$id_obra "  ;     // ejecuto DELETE
-       $Conn->query($sql) ;
+      if (!($Conn->query($sql)))   // si hay errores vamos a borrar una a una las UDOS 
+       {
+           $array_capitulos= DArray("id_capitulo", "Capitulos", "ID_OBRA='$id_obra'" ) ;
+           foreach ($array_capitulos as $id_capitulo) {
+                      $sql= "DELETE FROM `Capitulos`  WHERE ID_CAPITULO='$id_capitulo[0]'"  ;     // ejecuto DELETE
+                      $Conn->query($sql) ;
+           }
+
+       }    
 
        $sql= "DELETE FROM `SubObras`  WHERE ID_OBRA=$id_obra AND ID_SUBOBRA<>'$id_subobra_auto' "  ;     // ejecuto DELETE
-       $Conn->query($sql) ;
+      if (!($Conn->query($sql)))   // si hay errores vamos a borrar una a una las UDOS 
+       {
+           $array_subobra= DArray("ID_SUBOBRA", "SubObras", "ID_OBRA='$id_obra'   AND ID_SUBOBRA<>'$id_subobra_auto' " ) ;
+           foreach ($array_subobra as $id_subobra) {
+                      $sql= "DELETE FROM `SubObras`  WHERE ID_SUBOBRA='$id_subobra[0]'"  ;     // ejecuto DELETE
+                      $Conn->query($sql) ;
+           }
+
+       }    
     }
 }
 else
